Add dict and instance factory functions
When creating Model instances, even though fields may be null, they are
required with the None value.  These methods examine a Model and create
factory functions to use only the required fields to create either
complete Model dicts or Model instance.

Given default values are used when the fields are not provided.